Login page style in flexbox layout
