+++ react-app/src/index.css
@@ -1,1 +1,7 @@
 /* TODO Add site wide styles */
+html {
+  scroll-behavior: smooth;
+}
+body {
+  margin: 0;
+}
